+++ src/interfaces/IVolFeed.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.18;
 
 interface IVolFeed {
-  function getVol(uint128 strike, uint64 expiry) external view returns (uint128 vol, uint64 confidence);
+  function getVol(uint128 strike, uint64 expiry) external view returns (uint vol, uint confidence);
 
-  function getExpiryMinConfidence(uint64 expiry) external view returns (uint64 confidence);
+  function getExpiryMinConfidence(uint64 expiry) external view returns (uint confidence);
 
   ////////////
   // Events //
